nvram: ibm,skiboot NUL terminator check
NVRAM configuration strings are required to be NUL terminated and unused
data bytes in the partition should be set to NUL. Badly behaved system
software may not do this so same sanity checking is required. Ensuring
that the final data byte in a partition is a NUL should be sufficient.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>