console: keep existing stdout-path
Normally the OPAL console driver will set the linux,stdout-path to
so the Host OS will use the OPAL_*_CONSOLE calls to read and write
to the console. For platforms which pass through the UART directly
to the OS we want to set the stdout-path independently to the driver
so we can have the dummy console driver handle the OPAL calls sanely.

The existing dummy console driver will delete the pre-set
linux,stdout-path to handle because the Mambo simulator may pass in
a stdout-path. This patch also moves this hack into the mambo platform
init function where it belongs.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>