device: add dt_find_by_name_addr
Adds two helper functions that allow device nodes to be found via
their name and unit address. One will take an integer address and
format it to a hex string while the other looks up the unit based
a super supplied string. This is handy in a few places inside the
HDAT parser.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>