hw/phys_map: Use raw chip IDs
We need to be able to query the BAR mapping from the HDAT parser which
is run prior the the proc_chip structures being initialised. Define
a wrapper for the existing usages and provide a low-level function
that allows the map the be queried with the global hardware chip ID
itself.

Cc: Michael Neuling <mikey@neuling.org>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>