hdata_to_dt: Output up to PR_DEBUG
A lot of stuff that is useful for debugging and general sanity checking
of the HDAT parser is only printed at PR_DEBUG. Bump up the log level
for hdata_to_dt so that more of this is output by default. The actual
test cases only look at the DTS output so there's no harm in being
verbose.

Clean up the space indentation while we're here.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>