core/pcie-slots: Make dynamic slot creation generic
astbmc has some code to handle devices that are behind a "slot" on a
riser card that can't be added to the static slot tables for a system.
We probably want to use this code outside the slot table handling so
move it somewhere generic and rework it so slot table specifics aren't
buried inside it.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>