phb4: Make verbose EEH output less verbose
Each PHB4 component has a error status register and a set of
component-specific registers that describe that error in more detail. If
the status register of a component has no bits set there's not much
point in dumping the error detail registers.

This patch modifies the verbose EEH reporting logic so that it only
dumps the error registers of a component if there are bits set in the
error status register of that component. This limits the output to the
registers that are actually relevant to the error that occured which
makes interpreting the dumps slightly easier.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>