core/i2c: Make i2c_send_request() take an i2c_bus
We use OPAL bus IDs in the OPAL<->Host API, but there's no real reason
to use them inside of Skiboot. This patch makes i2c_send_request(), which
is internal to Skiboot, take a pointer to an i2c_bus structure rather an
OPAL I2C bus ID. This saves us from having to do a by_id lookup every
time we do an I2C request, simplifies the interface to the I2C core
and reduces the amount of runtime DT parsing, and reduces the amount of
error handling required when using this API.

This patch also converts the existing users (libstb and OpenCAPI)
which is where the bulk of the changes are.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>