opal-prd: Ignore the ibm, prefix on prd labels
Hostboot adds the ibm, prefix to the label of the various memory ranges that
it exports via HDAT. The prefix is hold-over from P8 where hostboot generated
the devicetree directly and needed to comply with the naming conventions.
HB would like to remove this, but opal-prd still has: "ibm,occ-common-area",
"ibm,homer-region", and "ibm,hbrt-code-image" hardcoded inside of it.

We can fix this by having opal-prd skip over the prefix when looking for a
given region.

Cc: Stable
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
I've left of fixing the internal names to keep the patch simply to
backport.

See https://github.com/open-power/boston-openpower/issues/675
for discussion.