hw/phb4: Add fake scom to allow reading PHB registers / IODA tables
The motivation for this is largely so that I can have some way to
extracting PHB state in a post-secureboot world. Currently the way
this is done is by wacking the PHB registers directly using /dev/mem,
but  that's pretty jank to begin with and with lockdown we can't do
that anyway.

What I figured we need is an architected way to access a register
space so this patch abuses to XSCOM API to do that. Each of the IODA
tables (and the PHB's HV regs) has a dedicated XSCOM "part id" which
can be used to access those registers from the host. Having a seperate
address space for the IODA tables is necessary since access them needs
to be done indirectly via the IODA_ADDR and IODA_DATA registers.

This setup works well enough for the non-config PHB registers and the
IODA tables since (like SCOMs) each register is 8 bytes. The cracks
start showing when you consider the root port config registers (part
of the HV register set) which only support four byte accesses. Moving
the config regs to a sepeate part ID would fix that, but eh.

The alternative would be something like adding a more generic set of
READ_REG / WRITE_REG OPAL calls which could handle things like varying
register access sizes, etc.

Comments welcome.

Not-Signed-off-by: Oliver O'Halloran <oohall@gmail.com>