test: Do gcov builds as a seperate pass
We only really use the gcov output when doing the coverage report as a
part of the "docs" CI builds. It's useful for development to just run
the unit tests so make sure the "check" and "coverage" targets are
seperate.

This also speeds up our CI builds since those jobs are already doing a
seperate GCOV pass so building and running the GCOV binaries during the
check pass is redundant.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
---
Tbh, we could probably drop the gcov builds outside the docs container.
We don't do anything with the gcov output for most of the builds, but I
guess it's useful to at least verify it builds.