travis: Disable seccomp on child containers
We've been hitting the following issue with valgrind breaking when run
inside a Fedora 32 container:

	        [ BOOT TEST ]  ccan/check_type/test/run
	valgrind: ccan/check_type/test/run: Operation not permitted
	make: *** [/build/ccan/Makefile.check:24:
	ccan/check_type/test/run-check] Error 126
	make: *** Waiting for unfinished jobs....

This seems to be a result of the default seccomp profile blocking system
calls that valgrind wants to use to inspect the process being debugged
(i.e. ptrace). Fix this by using the unrestricted seccomp profile and
giving the container the ptrace capability.

Travis uses ephemereal VMs for each job so this isn't any kind of
security issue.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>