travis: Add clang specific build job
Right now clang doesn't support the -ffixed-r<number> compiler options
that we rely on for the skiboot cpu_thread pointer. Remove the clang builds
from the debian-unstable and ubuntu-latest and builds in favour of a
clang specific job. This allows those jobs to pass normally and gives us
as specific job to monitor to see when support for those options lands
in clang.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>